+++ src/index.js
@@ -3,6 +3,7 @@ import standalone from "./standalone"
 import createPlugin from "./createPlugin"
 import rules from "./rules"
 import {
+  cssWordIsVariable, // Deprecated
   report,
   ruleMessages,
   styleSearch,
@@ -13,6 +14,7 @@ import createRuleTester from "./testUtils/createRuleTester"
 const stylelint = postcssPlugin
 
 stylelint.utils = {
+  cssWordIsVariable, // Deprecated
   report,
   ruleMessages,
   styleSearch,